refactor(design-lab): preserve review panel compatibility
